Extensive updates for FCS and other functionality: added new functions for getting parameters for FCS and FGCoefficients; added functionality for FGGain, FGSummer, FGGradient, and base class FGFCSComponent (Base Run() method now should almost always be called in the descendent class before running in order to set up Input and InputIdx

+++ FGAircraft.h
@@ -180,7 +180,7 @@ private:
   float WingArea, WingSpan, cbar;
   float Weight, EmptyWeight;
   float dt;
-  float CFGVersion;
+  double CFGVersion;
   string AircraftName;
   
   int numTanks;
